+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.27.0...master)
+### Added
+- Add period of record for each parameter in the monitoring location page parameter selection list.
+
 ## Fixed
 - Fixed bug that made compare time series disappear when the brush was adjusted by adding
 offsets to time series state in order to adjust x scale.